requirements.txt
pdfplumber
fastapi
uvicorn
python-multipart
groq
python-dotenv